@cverstege 2024-12-11
This also stops law from performing a DoS attack on remote storage systems :D

This could be improved further by also caching the files that are not present. For each file that is not present, still a gfal stat is performed. But this is already a good first improvement.