@cverstege 2024-12-11
The original problem is actually here: https://github.com/riga/law/blob/499d406ff29d3840daf6f2186905fbeca98f485b/law/workflow/remote.py#L359
Ideally, one would need to get rid of `self.task.as_branch(b).complete()` completely here to stop law from performing a stat on every file. Now, at least the present files in self._initially_existing_branches are skipped.