@cverstege 2024-12-11
I now implemented a way of disabling the explicit check. I just used it in one place so far, as I don't want to break other stuff.