@cverstege 2024-12-11
This PR requires #196 when using NestedSiblingFileCollections in remote workflows.