@riga 2024-12-15
Hi @cverstege ,

thanks for the deep dive into this performance bottleneck. I just quickly browsed over your PR and indeed, this additional loop over branch tasks, checking the existence of all branch outputs manually, is far from optimal.

Your changes look good, and I would even go one step further and fully cache branch values (keys) that are known to be complete over the entire progression of the workflow (i.e, without the need for the situation-dependent toggling of `skip_explicit_check`). Effectively, including the changes we made to the file collections recently, I think one could do this with only a single `collection.exists(keys=True)` (or `collection.count(keys=True)`) call.

I will have a more thorough look tomorrow and extend / generalize your suggestion.

---

**Sidenote:** This also aligns with observations that @mafrahm, @pkausw and I made in the past weeks. The scheduling phase of workflows, even large ones, used to be quite fast but since a while, this just takes far too long to be productive. I see now that this is very likely related to the issue you pointed out 🙏💯