@cverstege 2024-12-15
Thank you for looking into it in more detail. A similar behaviour also affects the scheduling of local workflows. Here, every task of the workflow is scheduled, also resulting in a check of every single output file. The information of a TargetCollection could therefore also be used when scheduling local workflows.