@riga 2024-12-15
A couple years ago, I added a [`DynamicRequirements` wrapper](https://github.com/spotify/luigi/blame/101b820dea0023f79092f5c17e096db9e89423a4/luigi/task.py#L778) to luigi for this exact purpose.

Maybe it's not being used in law in an optimal way, I will check next week.