@riga 2024-12-16
I just pushed 37592c9, which implements the full caching :+1:
Our workflows schedule quite fast now. Can you try as well?

I'm considering this PR superseded now, but we can still use it for discussions if there are some issues left.