@cverstege 2024-12-16
> A couple years ago, I added a [`DynamicRequirements` wrapper](https://github.com/spotify/luigi/blame/101b820dea0023f79092f5c17e096db9e89423a4/luigi/task.py#L778) to luigi for this exact purpose.
> 
> Maybe it's not being used in law in an optimal way, I will check next week.

Sorry for the noise, I think this acutally works. I just have the DEBUG output for the scheduler turned on. So i get spammed with stuff like this when running a local WF:
```
DEBUG: Checking if RunRivet(effective_workflow=local, branch=1988, campaign=LHC-NLO-ZplusJet_highpt, mc_setting=withNP, files_per_job=10, rivet_analyses=["ZplusJet_3", "ZplusJet_UE_ZFPT05ETA2P4", "MC_XS", "MC_WEIGHTS"], mc_generator=sherpa, number_of_gen_jobs=20000) is complete
INFO: Informed scheduler that task   RunRivet_1988__LHC_NLO_ZplusJet_ad68989d82   has status   DONE
```
But this has around the same waiting time as the fixed remote WFs. So I think that's working fine.