@cverstege 2024-12-16
I just tested job submission with the new master. This is really fast now. Thank you.