@riga 2024-12-17
Local workflow completion checks are now also accelerated (0f619e5) when sibling file collections are used :+1: